refactor(list): remove some circular dependencies for codesandbox
