+++ deno_dist/list/builder/block-builder.ts
@@ -1,19 +1,5 @@
-import { RimbuError } from '../../base/mod.ts';
 import type { OptLazy, TraverseState, Update } from '../../common/mod.ts';
-import type { BuilderBase, LeafBlock, NonLeafBlock } from '../list-custom.ts';
-
-export function createFromBlock<T>(
-  input: LeafBlock<T> | NonLeafBlock<T, any>
-): BlockBuilder<T, any> {
-  if (input.context.isLeafBlock(input)) {
-    return input.context.leafBlockBuilderSource(input);
-  }
-  if (input.context.isNonLeafBlock(input)) {
-    return input.context.nonLeafBlockBuilderSource(input);
-  }
-
-  RimbuError.throwInvalidStateError();
-}
+import type { BuilderBase } from '../list-custom.ts';
 
 export interface BlockBuilder<T, C = unknown> extends BuilderBase<T, C> {
   level: number;
