+++ packages/list/src/builder/nonleaf/tree-builder.ts
@@ -7,11 +7,7 @@ import type {
   NonLeafBuilder,
   NonLeafTree,
 } from '../../list-custom';
-import {
-  createFromBlock,
-  createNonLeaf,
-  TreeBuilderBase,
-} from '../../list-custom';
+import { TreeBuilderBase } from '../tree/tree-builder';
 
 export class NonLeafTreeBuilder<T, C extends BlockBuilder<T>>
   extends TreeBuilderBase<T, C>
@@ -31,12 +27,16 @@ export class NonLeafTreeBuilder<T, C extends BlockBuilder<T>>
 
   prepareMutate(): void {
     if (undefined !== this.source) {
-      this._left = createFromBlock(this.source.left) as any;
-      this._right = createFromBlock(this.source.right) as any;
+      this._left = this.context.createBlockBuilderFromBlock(
+        this.source.left
+      ) as any;
+      this._right = this.context.createBlockBuilderFromBlock(
+        this.source.right
+      ) as any;
       this._middle =
         null === this.source.middle
           ? undefined
-          : (createNonLeaf(this.source.middle) as any);
+          : (this.context.createNonLeafBuilder(this.source.middle) as any);
       this.source = undefined;
     }
   }
